improvement(falcon-server-env): moving endpoints to a separate class
