+++ packages/falcon-server-env/package.json
@@ -50,27 +50,30 @@
   },
   "dependencies": {
     "@deity/falcon-logger": "^0.0.3",
-    "apollo-datasource": "0.2.0",
-    "apollo-datasource-rest": "0.2.0",
+    "apollo-datasource": "0.2.1",
+    "apollo-datasource-rest": "0.2.1",
     "apollo-server-env": "2.2.0",
+    "apollo-server-caching": "0.2.1",
     "graphql": "14.0.2",
-    "html-entities": "^1.2.1",
-    "qs": "^6.5.2",
-    "striptags": "^3.1.1"
+    "html-entities": "1.2.1",
+    "qs": "6.6.0",
+    "striptags": "3.1.1"
   },
   "devDependencies": {
     "@types/graphql": "14.0.3",
     "@types/html-entities": "1.2.16",
     "@types/jest": "23.3.10",
+    "@types/koa": "2.0.47",
     "@types/koa-router": "7.0.35",
     "@types/nock": "9.3.0",
-    "@types/node": "10.12.10",
+    "@types/node": "10.12.12",
     "@types/qs": "6.5.1",
     "eventemitter2": "5.0.1",
+    "koa": "2.6.2",
     "koa-router": "7.4.0",
-    "nock": "10.0.1",
+    "nock": "10.0.3",
     "rimraf": "2.6.2",
-    "ts-jest": "23.10.4"
+    "ts-jest": "23.10.5"
   },
   "gitHead": "d8cd5eede81220aa29563275dcf35448611cc194"
 }
